+++ Models/InvoiceItem.cs
@@ -12,10 +12,10 @@ namespace invoice_system_backend.Models
         public string Description { get; set; }
 
         [Required]
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         [Required]
-        public float Price { get; set; }
+        public decimal Price { get; set; }
 
         [Required]
         [ForeignKey("Invoice")]
